backend/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
# djangorestframework-authtoken  # Use the latest stable or a specific version that exists
djangorestframework-simplejwt==4.7.2
psycopg2-binary==2.8.6
django-cors-headers==3.7.0
channels==3.0.3
channels-redis==3.2.0
daphne==3.0.2
redis==3.5.3